chore(aanbieder-view): reorganize
