+++ client/src/views/Authorized/Supplier/AanbiederView/AanbiederParticipantDetailView/AanbiederParticipantDetailOverviewView.tsx
@@ -1,26 +1,26 @@
+import React from 'react'
 import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
+
 import Headline, { SpacingType } from 'components/Chrome/Headline'
 import ErrorBlock from 'components/Core/Feedback/Error/ErrorBlock'
 import Spinner, { Animation } from 'components/Core/Feedback/Spinner/Spinner'
 import Center from 'components/Core/Layout/Center/Center'
 import Column from 'components/Core/Layout/Column/Column'
-import Paragraph from 'components/Core/Typography/Paragraph'
-import { AanbiederParticipantGoalDetailFields } from 'components/Domain/Aanbieder/AanbiederParticipantGoalDetailFields'
+import { AanbiederParticipantTab, AanbiederParticipantTabs } from 'components/Domain/Aanbieder/AanbiederParticipantTab'
 import { useMockQuery } from 'components/hooks/useMockQuery'
-import React from 'react'
-import { aanbiederParticipantDetail, AanbiederParticipantGoal } from '../../mocks'
+import { aanbiederParticipantDetail, AanbiederParticipantDetail } from '../mocks'
+import { AanbiederParticipantIntakeFields } from 'components/Domain/Aanbieder/AanbiederParticipantIntakeFields'
 
 interface Props {
     participantId: number
-    participantGoalId: number
 }
 
-export const AanbiederParticipantGoalDetailView: React.FunctionComponent<Props> = props => {
+export const AanbiederParticipantDetailOverviewView: React.FunctionComponent<Props> = ({ participantId }) => {
     const { i18n } = useLingui()
 
-    // TODO: replace with actual query
-    const { data, loading, error } = useMockQuery<AanbiederParticipantGoal>(aanbiederParticipantDetail.goals[0])
+    // TODO: replace with the api call/query (using participantId prop)
+    const { data, loading, error } = useMockQuery<AanbiederParticipantDetail>(aanbiederParticipantDetail)
 
     if (loading) {
         return (
@@ -33,9 +33,11 @@ export const AanbiederParticipantGoalDetailView: React.FunctionComponent<Props> 
     return (
         <>
             {/* TODO: add breadcrumb */}
-            <Paragraph bold={true}>{data?.participant.fullName || ''}</Paragraph>
-            <Headline spacingType={SpacingType.small} title={data?.name || ''} />
-            <Column spacing={10}>{renderList()}</Column>
+            <Headline spacingType={SpacingType.small} title={data?.fullName || ''} />
+            <Column spacing={10}>
+                <AanbiederParticipantTabs currentTab={AanbiederParticipantTab.overview} />
+                {renderList()}
+            </Column>
         </>
     )
 
@@ -49,6 +51,6 @@ export const AanbiederParticipantGoalDetailView: React.FunctionComponent<Props> 
             )
         }
 
-        return <AanbiederParticipantGoalDetailFields participantGoal={data} />
+        return <AanbiederParticipantIntakeFields participant={data} />
     }
 }
