+++ client/src/views/Authorized/Supplier/AanbiederView/AanbiederParticipantDetailView/AanbiederParticipantDetailView.tsx
@@ -2,14 +2,14 @@ import React from 'react'
 import { Route, Switch, useLocation } from 'react-router'
 import { routes } from 'routes/routes'
 import { AanbiederParticipantGoalsView } from './AanbiederParticipantGoalsView/AanbiederParticipantGoalsView'
-import { AanbiederParticipantOverviewView } from './AanbiederParticipantOverviewView'
+import { AanbiederParticipantDetailOverviewView } from './AanbiederParticipantDetailOverviewView'
 import { AanbiederParticipantRegistrationView } from './AanbiederParticipantRegistrationView'
 
 interface LocationStateProps {
     participantId: number
 }
 
-export const AanbiederParticipantView: React.FunctionComponent = () => {
+export const AanbiederParticipantDetailView: React.FunctionComponent = () => {
     const location = useLocation()
     const props = location.state as LocationStateProps
 
@@ -17,7 +17,7 @@ export const AanbiederParticipantView: React.FunctionComponent = () => {
 
     return (
         <Switch>
-            <Route path={basePath.overview} render={() => <AanbiederParticipantOverviewView {...props} />} />
+            <Route path={basePath.overview} render={() => <AanbiederParticipantDetailOverviewView {...props} />} />
             <Route path={basePath.registration} render={() => <AanbiederParticipantRegistrationView {...props} />} />
             <Route path={basePath.goals.index} component={AanbiederParticipantGoalsView} />
         </Switch>
