+++ client/src/views/Authorized/Supplier/AanbiederView/AanbiederParticipantDetailView/AanbiederParticipantGoalsView/AanbiederParticipantGoalsOverviewView.tsx
