+++ client/src/views/Authorized/Supplier/AanbiederView/AanbiederParticipantDetailView/AanbiederParticipantGoalsView/AanbiederParticipantGoalsView.tsx
